legacy/streamlit/web/requirements.txt
streamlit>=1.36
pyyaml>=6.0
feedparser>=6.0
trafilatura>=1.6
ebooklib>=0.18
Pillow>=10.0
google-api-python-client>=2.100
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
requests>=2.31
